third_party/blink/web_tests/virtual/customizable-select-disabled/html5lib/generated/run-tests2-write-expected.txt

CONSOLE ERROR: A <select> tag was parsed within another <select> tag and was converted into </select>. This behavior will change in a future browser version. Please add the missing </select> end tag.
../resources/tests2.dat: PASS
